+++ src/nl_fit/ceres.rs
@@ -1,6 +1,7 @@
 use crate::nl_fit::constants::PARAMETER_TOLERANCE;
 use crate::nl_fit::curve_fit::{CurveFitResult, CurveFitTrait};
 use crate::nl_fit::data::Data;
+use crate::nl_fit::prior::ln_prior::LnPriorEvaluator;
 
 use ceres_solver::{CurveFitProblem1D, CurveFunctionType, LossFunction, SolverOptions};
 use ndarray::Zip;
@@ -71,7 +72,7 @@ impl CurveFitTrait for CeresCurveFit {
     where
         F: 'static + Clone + Fn(f64, &[f64; NPARAMS]) -> f64,
         DF: 'static + Clone + Fn(f64, &[f64; NPARAMS], &mut [f64; NPARAMS]),
-        LP: Clone + Fn(&[f64; NPARAMS]) -> f64,
+        LP: LnPriorEvaluator<NPARAMS>,
     {
         let func: CurveFunctionType = {
             let model = model.clone();
@@ -162,10 +163,6 @@ mod tests {
         derivatives[2] = 1.0;
     }
 
-    fn nonlinear_func_dump_ln_prior(_param: &[f64; 3]) -> f64 {
-        0.0
-    }
-
     #[test]
     fn nonlinear() {
         const N: usize = 300;
@@ -197,7 +194,7 @@ mod tests {
             (&[f64::NEG_INFINITY; 3], &[f64::INFINITY; 3]),
             nonlinear_func,
             nonlinear_func_derivatives,
-            nonlinear_func_dump_ln_prior,
+            crate::nl_fit::LnPrior::none(),
         );
 
         // curve_fit(lambda x, a, b, c: b * np.exp(-a * x) * x**2 + c, xdata=t, ydata=y, sigma=1/np.array(inv_err), p0=[1, 1, 1], xtol=1e-6)