+++ src/nl_fit/curve_fit.rs
@@ -6,6 +6,7 @@ use crate::nl_fit::lmsder::LmsderCurveFit;
 use crate::nl_fit::mcmc::McmcCurveFit;
 #[cfg(feature = "nuts")]
 use crate::nl_fit::nuts::NutsCurveFit;
+use crate::nl_fit::prior::ln_prior::LnPriorEvaluator;
 
 use enum_dispatch::enum_dispatch;
 use schemars::JsonSchema;
@@ -35,7 +36,7 @@ pub trait CurveFitTrait: Clone + Debug + Serialize + DeserializeOwned {
     where
         F: 'static + Clone + Fn(f64, &[f64; NPARAMS]) -> f64,
         DF: 'static + Clone + Fn(f64, &[f64; NPARAMS], &mut [f64; NPARAMS]),
-        LP: Clone + Fn(&[f64; NPARAMS]) -> f64;
+        LP: LnPriorEvaluator<NPARAMS>;
 }
 
 /// Optimization algorithm for non-linear least squares